// src/interfaces/Card.ts
export default interface Card {
    id?: number;
    userId?: number;
    title?: string;
    subTitle?: string;
    description?: string;
    web?: string;
    imageUrl?: string;
    imageAlt?: string;
    phone?: string;
    email?: string;
    state?: string;
    country?: string;
    city?: string;
    street?: string;
    houseNumber?: string;
    zip?: number;
}